requirements.txt
ultralytics>=8.3.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
albumentations>=1.4.0
torch>=2.0.0
torchvision>=0.15.0
grad-cam>=1.5.0
imagehash>=4.3.1
Pillow>=10.0.0
roboflow>=1.1.0
